Move integration API to `me.snoty.integration`
