+++ api/src/main/kotlin/me/snoty/backend/User.kt
@@ -4,9 +4,6 @@ import java.util.*
 import kotlinx.serialization.Serializable
 import me.snoty.backend.utils.UUIDSerializer
 
-/*
-@Serializable
-*/
 @Serializable
 data class User(
 	@Serializable(with = UUIDSerializer::class)
