--- integrations/api/src/main/kotlin/me/snoty/backend/User.kt
@@ -1,1 +1,0 @@
-package me.snoty.backend
