+++ integrations/api/src/main/kotlin/me/snoty/integration/common/Fetcher.kt
@@ -1,6 +1,6 @@
-package me.snoty.backend.integration.common
+package me.snoty.integration.common
 
-import me.snoty.backend.integration.common.diff.EntityDiffMetrics
+import me.snoty.integration.common.diff.EntityDiffMetrics
 import me.snoty.backend.scheduling.JobRequest
 import me.snoty.backend.scheduling.JobRequestHandler
