+++ integrations/api/src/main/kotlin/me/snoty/integration/common/InstanceId.kt
@@ -1,4 +1,4 @@
-package me.snoty.backend.integration.common
+package me.snoty.integration.common
 
 /**
  * Unique ID representing a 3rd party service instance.
