+++ integrations/api/src/main/kotlin/me/snoty/integration/common/Integration.kt
@@ -1,4 +1,4 @@
-package me.snoty.backend.integration.common
+package me.snoty.integration.common
 
 import me.snoty.backend.User
 import kotlin.reflect.KClass
