+++ integrations/api/src/main/kotlin/me/snoty/integration/common/IntegrationScheduler.kt
@@ -1,4 +1,4 @@
-package me.snoty.backend.integration.common
+package me.snoty.integration.common
 
 import me.snoty.backend.scheduling.JobRequest
 import me.snoty.backend.scheduling.Scheduler
