+++ integrations/api/src/main/kotlin/me/snoty/integration/common/IntegrationSettings.kt
@@ -1,4 +1,4 @@
-package me.snoty.backend.integration.common
+package me.snoty.integration.common
 
 import kotlinx.serialization.Serializable
