+++ integrations/api/src/main/kotlin/me/snoty/integration/common/diff/EntityDiffMetrics.kt
@@ -1,4 +1,4 @@
-package me.snoty.backend.integration.common.diff
+package me.snoty.integration.common.diff
 
 import io.micrometer.core.instrument.Counter
 import io.micrometer.core.instrument.MeterRegistry
