+++ integrations/webuntis/src/main/resources/META-INF/services/me.snoty.integration.common.IntegrationFactory
